Currency costs for teleporting to a home within multiple zones is now cumulative
The currency costs for teleporting to a particular home in multiple zones is now cumulative. Costs per currency are then set to the sum of the cost of that currency for teleporting to the relevant zones or the cost of that currency for teleporting to that world, whichever is higher.